source/rule-engine/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
stomp.py